renames card-details.jsx to details-page
