+++ src/pages/details-page.jsx
